Moves postFavorites function into post-handlers file
