+++ lib/index.js
@@ -2,8 +2,6 @@
 import './styles.scss';
 import * as errors from './errors';
 import * as post from './post-handlers';
-import CurrentWeather from './current-weather'
-
 
 $(document).ready(function() {
   renderFavorites();
@@ -90,20 +88,7 @@ const renderFavorites = () => {
           <div class="favorites">  
           </div>
         </div>`);
-    postFavorites();
+    post.postFavorites();
   };
 }
 
-const postFavorites = () => {
-  const favoritesUrl = API_URL + "/favorites"
-  const api_key = document.cookie.split('=')[1]
-  $.get(favoritesUrl, { api_key })
-    .then(function(data) {
-      const favorites = data.data.map(favorite => {
-        return new CurrentWeather(favorite)
-      });
-      favorites.forEach(favorite => {
-        post.postFavorite(favorite)
-      })
-    })
-}
