+++ lib/post-handlers.js
@@ -70,6 +70,20 @@ export const postWeatherDays = (weatherDays) => {
   })
 };
 
+export const postFavorites = () => {
+  const favoritesUrl = API_URL + "/favorites"
+  const api_key = document.cookie.split('=')[1]
+  $.get(favoritesUrl, { api_key })
+    .then(function(data) {
+      const favorites = data.data.map(favorite => {
+        return new CurrentWeather(favorite)
+      });
+      favorites.forEach(favorite => {
+        postFavorite(favorite)
+      })
+    })
+};
+
 export const postFavorite = (favorite) => {
   if(!favorite) {
     favorite = JSON.parse(localStorage.getItem("currentWeather"));
